refactor/react: consolidate and extend React hooks functionality
- Merge similar hooks into unified implementations
- Add support for immutable state and models
- Introduce new hooks for exception handling and form management
- Update existing hooks to provide more comprehensive functionality
- Refactor code to improve readability and maintainability
