+++ packages/react/src/array.ts
@@ -1,5 +1,6 @@
-import { type LinkableSchema, type StateOptions } from '@anchor/core';
-import { useAnchor } from './anchor.js';
+import { anchor, type LinkableSchema, type StateOptions } from '@anchor/core';
+import { useVariable } from './ref.js';
+import type { AnchorState } from './types.js';
 
 /**
  * Creates a reactive ordered list state using Anchor's state management.
@@ -24,6 +25,33 @@ export function useOrderedList<T extends unknown[], S extends LinkableSchema = L
   init: T,
   compare: (a: T[number], b: T[number]) => number,
   options?: StateOptions<S>
-) {
-  return useAnchor(init, { ...options, ordered: true, compare });
+): AnchorState<T> {
+  const [state, setState] = useVariable(() => {
+    return anchor.ordered(init, compare, options);
+  }, [init, options]);
+  return [state.value, state, setState];
+}
+
+/**
+ * Creates a reactive flat list state using Anchor's state management.
+ *
+ * This hook provides a way to manage an array state that only reacts to array mutations while maintain the
+ * reactivity recursively.
+ *
+ * @template T - The type of the array elements
+ * @template S - The schema type for linkable state (defaults to LinkableSchema)
+ *
+ * @param init - The initial array state
+ * @param options - Optional state configuration options
+ *
+ * @returns A reactive state object with methods to interact with the flat list
+ */
+export function useFlatList<T extends unknown[], S extends LinkableSchema = LinkableSchema>(
+  init: T,
+  options?: StateOptions<S>
+): AnchorState<T> {
+  const [state, setState] = useVariable(() => {
+    return anchor.flat(init, options);
+  }, [init, options]);
+  return [state.value, state, setState];
 }
